+++ Engine/Engine.vcxproj
@@ -43,13 +43,10 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Camera.cpp" />
+    <ClCompile Include="cameras\Camera.cpp" />
+    <ClCompile Include="cameras\Player.cpp" />
     <ClCompile Include="Engine.cpp" />
-    <ClCompile Include="GameObject.cpp" />
     <ClCompile Include="Logger.cpp" />
-    <ClCompile Include="Material.cpp" />
-    <ClCompile Include="Mesh.cpp" />
-    <ClCompile Include="Player.cpp" />
     <ClCompile Include="renderer\Renderer.cpp" />
     <ClCompile Include="renderer\VulkanBuffer.cpp" />
     <ClCompile Include="renderer\VulkanCommandBuffer.cpp" />
@@ -60,23 +57,24 @@
     <ClCompile Include="renderer\VulkanPipeline.cpp" />
     <ClCompile Include="renderer\VulkanRenderPass.cpp" />
     <ClCompile Include="renderer\VulkanSwapchain.cpp" />
-    <ClCompile Include="Scene.cpp" />
     <ClCompile Include="Source.cpp" />
-    <ClCompile Include="Texture.cpp" />
     <ClCompile Include="Window.cpp" />
+    <ClCompile Include="world\GameObject.cpp" />
+    <ClCompile Include="world\Material.cpp" />
+    <ClCompile Include="world\Mesh.cpp" />
+    <ClCompile Include="world\Scene.cpp" />
+    <ClCompile Include="world\Texture.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="Camera.h" />
-    <ClInclude Include="Config.h" />
+    <ClInclude Include="cameras\Camera.h" />
+    <ClInclude Include="cameras\Player.h" />
     <ClInclude Include="Engine.h" />
     <ClInclude Include="errors.h" />
-    <ClInclude Include="GameObject.h" />
-    <ClInclude Include="Graphics.h" />
+    <ClInclude Include="graphics\Config.h" />
+    <ClInclude Include="graphics\Graphics.h" />
+    <ClInclude Include="graphics\Math.h" />
+    <ClInclude Include="graphics\Vertex.h" />
     <ClInclude Include="Logger.h" />
-    <ClInclude Include="Material.h" />
-    <ClInclude Include="Math.h" />
-    <ClInclude Include="Mesh.h" />
-    <ClInclude Include="Player.h" />
     <ClInclude Include="renderer\Renderer.h" />
     <ClInclude Include="renderer\VulkanBuffer.h" />
     <ClInclude Include="renderer\VulkanCommandBuffer.h" />
@@ -87,10 +85,12 @@
     <ClInclude Include="renderer\VulkanPipeline.h" />
     <ClInclude Include="renderer\VulkanRenderPass.h" />
     <ClInclude Include="renderer\VulkanSwapchain.h" />
-    <ClInclude Include="Scene.h" />
-    <ClInclude Include="Texture.h" />
-    <ClInclude Include="Vertex.h" />
     <ClInclude Include="Window.h" />
+    <ClInclude Include="world\GameObject.h" />
+    <ClInclude Include="world\Material.h" />
+    <ClInclude Include="world\Mesh.h" />
+    <ClInclude Include="world\Scene.h" />
+    <ClInclude Include="world\Texture.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="..\shaders\shader.frag" />
