+++ Engine/Engine.vcxproj.filters
@@ -25,11 +25,11 @@
     <Filter Include="Engine\Graphics">
       <UniqueIdentifier>{c8467a9d-cccc-486f-814e-ac1fa2cce737}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Engine\Graphics\Header">
-      <UniqueIdentifier>{0773a226-0a6e-44bb-997b-0848af431efa}</UniqueIdentifier>
+    <Filter Include="Engine\Cameras">
+      <UniqueIdentifier>{3bcbc072-07c8-49df-926a-7e4a2a3da833}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Engine\Graphics\Source">
-      <UniqueIdentifier>{2e3255ef-0cc2-44cf-9e70-64ccc5b203bc}</UniqueIdentifier>
+    <Filter Include="Engine\World">
+      <UniqueIdentifier>{f0d3173b-0d3b-48a2-a5fb-0ec342f99870}</UniqueIdentifier>
     </Filter>
   </ItemGroup>
   <ItemGroup>
@@ -38,33 +38,6 @@
     </ClCompile>
     <ClCompile Include="Logger.cpp">
       <Filter>Source</Filter>
-    </ClCompile>
-    <ClCompile Include="Engine.cpp">
-      <Filter>Engine</Filter>
-    </ClCompile>
-    <ClCompile Include="Camera.cpp">
-      <Filter>Engine\Graphics\Source</Filter>
-    </ClCompile>
-    <ClCompile Include="GameObject.cpp">
-      <Filter>Engine\Graphics\Source</Filter>
-    </ClCompile>
-    <ClCompile Include="Material.cpp">
-      <Filter>Engine\Graphics\Source</Filter>
-    </ClCompile>
-    <ClCompile Include="Mesh.cpp">
-      <Filter>Engine\Graphics\Source</Filter>
-    </ClCompile>
-    <ClCompile Include="Player.cpp">
-      <Filter>Engine\Graphics\Source</Filter>
-    </ClCompile>
-    <ClCompile Include="Scene.cpp">
-      <Filter>Engine\Graphics\Source</Filter>
-    </ClCompile>
-    <ClCompile Include="Texture.cpp">
-      <Filter>Engine\Graphics\Source</Filter>
-    </ClCompile>
-    <ClCompile Include="Window.cpp">
-      <Filter>Engine\Graphics\Source</Filter>
     </ClCompile>
     <ClCompile Include="renderer\Renderer.cpp">
       <Filter>Engine\Renderer\Source</Filter>
@@ -96,49 +69,37 @@
     <ClCompile Include="renderer\VulkanSwapchain.cpp">
       <Filter>Engine\Renderer\Source</Filter>
     </ClCompile>
+    <ClCompile Include="cameras\Camera.cpp">
+      <Filter>Engine\Cameras</Filter>
+    </ClCompile>
+    <ClCompile Include="world\GameObject.cpp">
+      <Filter>Engine\World</Filter>
+    </ClCompile>
+    <ClCompile Include="world\Material.cpp">
+      <Filter>Engine\World</Filter>
+    </ClCompile>
+    <ClCompile Include="world\Mesh.cpp">
+      <Filter>Engine\World</Filter>
+    </ClCompile>
+    <ClCompile Include="world\Scene.cpp">
+      <Filter>Engine\World</Filter>
+    </ClCompile>
+    <ClCompile Include="world\Texture.cpp">
+      <Filter>Engine\World</Filter>
+    </ClCompile>
+    <ClCompile Include="cameras\Player.cpp">
+      <Filter>Engine\Cameras</Filter>
+    </ClCompile>
+    <ClCompile Include="Window.cpp">
+      <Filter>Source</Filter>
+    </ClCompile>
+    <ClCompile Include="Engine.cpp">
+      <Filter>Source</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="Logger.h">
       <Filter>Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Vertex.h">
-      <Filter>Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Math.h">
-      <Filter>Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Engine.h">
-      <Filter>Engine</Filter>
-    </ClInclude>
-    <ClInclude Include="Graphics.h">
-      <Filter>Engine\Renderer\Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Camera.h">
-      <Filter>Engine\Graphics\Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Config.h">
-      <Filter>Engine\Graphics\Header</Filter>
-    </ClInclude>
-    <ClInclude Include="GameObject.h">
-      <Filter>Engine\Graphics\Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Material.h">
-      <Filter>Engine\Graphics\Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Mesh.h">
-      <Filter>Engine\Graphics\Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Player.h">
-      <Filter>Engine\Graphics\Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Scene.h">
-      <Filter>Engine\Graphics\Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Texture.h">
-      <Filter>Engine\Graphics\Header</Filter>
-    </ClInclude>
-    <ClInclude Include="Window.h">
-      <Filter>Engine\Graphics\Header</Filter>
     </ClInclude>
     <ClInclude Include="renderer\Renderer.h">
       <Filter>Engine\Renderer\Header</Filter>
@@ -171,6 +132,45 @@
       <Filter>Engine\Renderer\Header</Filter>
     </ClInclude>
     <ClInclude Include="errors.h">
+      <Filter>Header</Filter>
+    </ClInclude>
+    <ClInclude Include="cameras\Camera.h">
+      <Filter>Engine\Cameras</Filter>
+    </ClInclude>
+    <ClInclude Include="world\GameObject.h">
+      <Filter>Engine\World</Filter>
+    </ClInclude>
+    <ClInclude Include="world\Material.h">
+      <Filter>Engine\World</Filter>
+    </ClInclude>
+    <ClInclude Include="world\Mesh.h">
+      <Filter>Engine\World</Filter>
+    </ClInclude>
+    <ClInclude Include="world\Scene.h">
+      <Filter>Engine\World</Filter>
+    </ClInclude>
+    <ClInclude Include="world\Texture.h">
+      <Filter>Engine\World</Filter>
+    </ClInclude>
+    <ClInclude Include="cameras\Player.h">
+      <Filter>Engine\Cameras</Filter>
+    </ClInclude>
+    <ClInclude Include="graphics\Vertex.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="graphics\Graphics.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="graphics\Config.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="graphics\Math.h">
+      <Filter>Engine\Graphics</Filter>
+    </ClInclude>
+    <ClInclude Include="Window.h">
+      <Filter>Header</Filter>
+    </ClInclude>
+    <ClInclude Include="Engine.h">
       <Filter>Header</Filter>
     </ClInclude>
   </ItemGroup>
