+++ Engine/world/Mesh.cpp
@@ -3,7 +3,7 @@
 #ifndef TINYOBJLOADER_IMPLEMENTATION
 
 #define TINYOBJLOADER_IMPLEMENTATION
-#include "tiny_obj_loader.h"
+#include "../tiny_obj_loader.h"
 
 #endif // !TINYOBJLOADER_IMPLEMENTATION
